requirements.txt
rejson
jsonpath-extractor